requirements.txt
pygame>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.0.0
librosa>=0.10.0
pydub>=0.25.1
requests>=2.31.0
